+++ functions/get-adgroupuser.ps1
@@ -1,10 +1,8 @@
-
-
 Function Get-ADGroupUser {
     [cmdletbinding()]
     [OutputType("ADGroupUser")]
     Param(
-        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the name of an Active Directory group.",ValueFromPipelineByPropertyName,ValueFromPipeline)]
+        [Parameter(Position = 0, Mandatory, HelpMessage = "Enter the name of an Active Directory group.", ValueFromPipelineByPropertyName, ValueFromPipeline)]
         [string]$Name,
         [Parameter(HelpMessage = "Specify a domain controller to query.")]
         [alias("dc", "domaincontroller")]
@@ -29,32 +27,48 @@ Function Get-ADGroupUser {
 
     Process {
         Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Getting members of $Name"
-        $group= Get-ADGroup -identity  "$($Name)"
+        $group = Get-ADGroup -Identity "$($Name)"
         if ($group) {
-        $members = $group   |  Get-ADGroupMember  -Recursive
+            $members = $group | Get-ADGroupMember -Recursive
 
-        Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($members.count) members in total."
-        #get details about each user
-        foreach ($member in $members) {
-            $user = Get-ADUser -Identity $member.distinguishedname -Properties PasswordLastSet, Description, Title, Displayname, Department
-            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($user.distinguishedname)"
-            [pscustomobject]@{
-                PSTypeName        = "ADGroupUser"
-                DistinguishedName = $user.DistinguishedName
-                Displayname       = $user.Displayname
-                Name              = $user.name
-                Title             = $user.title
-                Department        = $user.Department
-                Description       = $user.Description
-                Enabled           = $user.Enabled
-                PasswordLastSet   = $user.PasswordLastSet
-                Group = $group.DistinguishedName
+            Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Found $($members.count) members in total."
+            #get details about each group member
+            $getParams = @{
+                properties = 'PasswordLastSet', 'Description', ' Title', 'Displayname', 'Department'
+                Identity   = ""
             }
-        } #foreach
-    } #if group was found
-    else {
-        Write-Warning "Failed to find an Active Directory group called $Name"
-    }
+            foreach ($member in $members) {
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] member is a $($member.objectclass)"
+                $getParams.Identity = $member.distinguishedname
+
+                <#
+                    Not using Get-ADObject because it can't return the passwordLastSet property
+                #>
+                if ($member.objectclass -eq "user") {
+                    $grpMember = Get-ADUser @getparams
+                }
+                elseif ($member.objectclass -eq "computer") {
+                    $grpMember = Get-ADComputer @getparams
+                }
+   
+                Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] $($grpMember.distinguishedname)"
+                [pscustomobject]@{
+                    PSTypeName        = "ADGroupUser"
+                    DistinguishedName = $grpMember.DistinguishedName
+                    Displayname       = $grpMember.Displayname
+                    Name              = $grpMember.name
+                    Title             = $grpMember.title
+                    Department        = $grpMember.Department
+                    Description       = $grpMember.Description
+                    Enabled           = $grpMember.Enabled
+                    PasswordLastSet   = $grpMember.PasswordLastSet
+                    Group             = $group.DistinguishedName
+                }
+            } #foreach
+        } #if group was found
+        else {
+            Write-Warning "Failed to find an Active Directory group called $Name"
+        }
     } #process
 
     End {